concurrency: move config in other file
